Moves modularity into quality module.
