+++ networkx/algorithms/community/__init__.py
@@ -3,5 +3,4 @@ from networkx.algorithms.community.centrality import *
 from networkx.algorithms.community.community_generators import *
 from networkx.algorithms.community.kclique import *
 from networkx.algorithms.community.kernighan_lin import *
-from networkx.algorithms.community.modularity import *
 from networkx.algorithms.community.quality import *
