--- networkx/algorithms/community/modularity.py
@@ -1,93 +1,0 @@
-# modularity.py - functions for computing modularity of a partition
-#
-# Copyright 2011 Ben Edwards <bedwards@cs.unm.edu>.
-# Copyright 2011 Aric Hagberg <hagberg@lanl.gov>.
-# Copyright 2015 NetworkX developers.
-#
-# This file is part of NetworkX.
-#
-# NetworkX is distributed under a BSD license; see LICENSE.txt for more
-# information.
-"""Functions for computing the modularity of a partition."""
-from __future__ import division
-
-from itertools import product
-
-import networkx as nx
-from .community_utils import is_partition
-
-
-def modularity(G, communities, weight='weight'):
-    r"""Returns the modularity of the given partition of the graph.
-
-    Modularity is defined in [1] as
-
-    .. math::
-
-        Q = \frac{1}{2m} \sum_{ij} \left( A_{ij} - \frac{k_ik_j}{2m}\right)
-            \delta(c_i,c_j)
-
-    where `m` is the number of edges, `A` is the adjacency matrix of ``G``,
-    `k_i` is the degree of `i` and `\delta(c_i,c_j)` is 1 if `i` and `j`
-    are in the same community and 0 otherwise.
-
-    Parameters
-    ----------
-    G : NetworkX Graph
-
-    communities : list
-        List of sets of nodes of ``G`` representing a partition of the
-        nodes.
-
-    Returns
-    -------
-    Q : float
-        The modularity of the paritition.
-
-    Raises
-    ------
-    NetworkXError
-        If ``communities`` is not a partition of the nodes of ``G``.
-
-    Examples
-    --------
-    >>> G = nx.barbell_graph(3, 0)
-    >>> nx.modularity(G, [{0, 1, 2}, {3, 4, 5}])
-    0.35714285714285704
-
-    References
-    ----------
-    .. [1] M. E. J. Newman *Networks: An Introduction*, page 224.
-       Oxford University Press, 2011.
-
-    """
-    if not is_partition(G, communities):
-        raise nx.NetworkXError('communities is not a partition of G')
-
-    multigraph = G.is_multigraph()
-    directed = G.is_directed()
-    m = G.size(weight=weight)
-    if directed:
-        out_degree = dict(G.out_degree(weight=weight))
-        in_degree = dict(G.in_degree(weight=weight))
-        norm = 1 / m
-    else:
-        out_degree = dict(G.degree(weight=weight))
-        in_degree = out_degree
-        norm = 1 / (2 * m)
-
-    def val(u, v):
-        try:
-            if multigraph:
-                w = sum(d.get(weight, 1) for k, d in G[u][v].items())
-            else:
-                w = G[u][v].get(weight, 1)
-        except KeyError:
-            w = 0
-        # Double count self-loops if the graph is undirected.
-        if u == v and not directed:
-            w *= 2
-        return w - in_degree[u] * out_degree[v] * norm
-
-    Q = sum(val(u, v) for c in communities for u, v in product(c, repeat=2))
-    return Q * norm
